finance_fusion/requirement.txt
Django==5.1.5
psycopg2-binary==2.9.8
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
python-decouple==3.8